requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.2
altair==5.0.1
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asyncio==3.4.3
asyncssh==2.13.2
asynctest==0.13.0
atpublic==3.1.2
attrs==23.1.0
autocommand==2.2.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
configobj==5.0.8
cryptography==41.0.3
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
docker==6.1.3
docopt==0.6.2
dpath==2.1.6
dulwich==0.21.5
dvc==2.10.2
dvc-render==0.0.5
dvclive==0.9.0
entrypoints==0.4
exceptiongroup==1.1.2
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.2.5
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
funcy==2.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
grandalf==0.6
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.2
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.12.0
inflect==6.0.5
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.collections==4.2.0
jaraco.context==4.3.0
jaraco.functools==3.7.0
jaraco.structures==2.1.0
jaraco.text==3.11.1
jaraco.ui==2.3.0
jaraco.windows==5.7.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mlflow==1.30.1
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
nanotime==0.5.2
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.7
networkx==2.6.3
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
optuna==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
plotly==5.16.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
pydot==1.4.2
pygit2==1.10.1
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-benedict==0.32.0
python-box==7.0.1
python-dateutil==2.8.2
python-fsutil==0.10.0
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
ray==2.6.1
requests==2.31.0
retrying==1.3.4
rich==13.5.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==0.0.19
seaborn==0.12.2
Send2Trash==1.8.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
streamlit==1.23.1
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torch-geometric==2.3.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
voluptuous==0.13.1
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yarg==0.1.9
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.15.0
